// src/modules/notifications/enums/notification-type.enum.ts
export enum NotificationType {
  TASK_ASSIGNED = 'TASK_ASSIGNED',
  TASK_STATUS_CHANGED = 'TASK_STATUS_CHANGED',
  TASK_COMMENTED = 'TASK_COMMENTED',
  TASK_SUBMITTED_FOR_REVIEW = 'TASK_SUBMITTED_FOR_REVIEW',
  TASK_APPROVED = 'TASK_APPROVED',
  TASK_CHANGES_REQUESTED = 'TASK_CHANGES_REQUESTED',
  REVIEW_WAITING_24H = 'REVIEW_WAITING_24H',
  REVIEW_WAITING_48H = 'REVIEW_WAITING_48H',

  POST_SUBMITTED_TO_CLIENT = 'POST_SUBMITTED_TO_CLIENT',
  POST_CHANGES_REQUESTED = 'POST_CHANGES_REQUESTED',
  POST_APPROVED = 'POST_APPROVED',
  POST_PUBLISHED = 'POST_PUBLISHED',
  POST_COMMENTED = 'POST_COMMENTED',
  POST_REJECTED = 'POST_REJECTED',

  LEAD_ASSIGNED = 'LEAD_ASSIGNED',
  LEAD_STATUS_CHANGED = 'LEAD_STATUS_CHANGED',
  LEAD_NOTE_ADDED = 'LEAD_NOTE_ADDED',
  INVITATION_CREATED = 'INVITATION_CREATED',
  INVITATION_ACCEPTED = 'INVITATION_ACCEPTED',

  REPORT_CREATED = 'REPORT_CREATED',
}