// src/modules/tasks/enums/task-activity-action.enum.ts
export enum TaskActivityAction {
  CREATED = 'CREATED',
  UPDATED = 'UPDATED',
  STATUS_CHANGED = 'STATUS_CHANGED',
  COMMENTED = 'COMMENTED',
  ATTACHMENT_ADDED = 'ATTACHMENT_ADDED',
  ATTACHMENT_REMOVED = 'ATTACHMENT_REMOVED',
  TASK_SUBMITTED_FOR_REVIEW = 'TASK_SUBMITTED_FOR_REVIEW',
  TASK_APPROVED = 'TASK_APPROVED',
  TASK_CHANGES_REQUESTED = 'TASK_CHANGES_REQUESTED',
  TASK_APPROVER_CHANGED = 'TASK_APPROVER_CHANGED',
}